fix: ajuste das propriedades nas entidades (PIX, TED e Trasacao)
